Add Concerns namespace and refactor code to reuse traits for rules (#240)
* Add Concerns namespace and refactor code to reuse traits for rules

* Create profileRules method for trait
