+++ app/Concerns/PasswordValidationRules.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Actions\Fortify;
+namespace App\Concerns;
 
 use Illuminate\Validation\Rules\Password;
 
@@ -14,5 +14,15 @@ trait PasswordValidationRules
     protected function passwordRules(): array
     {
         return ['required', 'string', Password::default(), 'confirmed'];
+    }
+
+    /**
+     * Get the validation rules used to validate the current password.
+     *
+     * @return array<int, \Illuminate\Contracts\Validation\Rule|array<mixed>|string>
+     */
+    protected function currentPasswordRules(): array
+    {
+        return ['required', 'string', 'current_password'];
     }
 }
