+++ app/Http/Requests/Settings/PasswordUpdateRequest.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Requests\Settings;
 
+use App\Concerns\PasswordValidationRules;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Validation\Rules\Password;
 
 class PasswordUpdateRequest extends FormRequest
 {
+    use PasswordValidationRules;
+
     /**
      * Get the validation rules that apply to the request.
      *
@@ -16,8 +18,8 @@ class PasswordUpdateRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'current_password' => ['required', 'current_password'],
-            'password' => ['required', Password::defaults(), 'confirmed'],
+            'current_password' => $this->currentPasswordRules(),
+            'password' => $this->passwordRules(),
         ];
     }
 }
