+++ app/Http/Requests/Settings/ProfileDeleteRequest.php
@@ -2,10 +2,13 @@
 
 namespace App\Http\Requests\Settings;
 
+use App\Concerns\PasswordValidationRules;
 use Illuminate\Foundation\Http\FormRequest;
 
 class ProfileDeleteRequest extends FormRequest
 {
+    use PasswordValidationRules;
+
     /**
      * Get the validation rules that apply to the request.
      *
@@ -14,7 +17,7 @@ class ProfileDeleteRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'password' => ['required', 'current_password'],
+            'password' => $this->currentPasswordRules(),
         ];
     }
 }
