+++ app/Http/Requests/Settings/ProfileUpdateRequest.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Requests\Settings;
 
-use App\Models\User;
+use App\Concerns\ProfileValidationRules;
 use Illuminate\Contracts\Validation\ValidationRule;
 use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Validation\Rule;
 
 class ProfileUpdateRequest extends FormRequest
 {
+    use ProfileValidationRules;
+
     /**
      * Get the validation rules that apply to the request.
      *
@@ -16,16 +17,6 @@ class ProfileUpdateRequest extends FormRequest
      */
     public function rules(): array
     {
-        return [
-            'name' => ['required', 'string', 'max:255'],
-            'email' => [
-                'required',
-                'string',
-                'lowercase',
-                'email',
-                'max:255',
-                Rule::unique(User::class)->ignore($this->user()->id),
-            ],
-        ];
+        return $this->profileRules($this->user()->id);
     }
 }
